halves communicate now, also keyball communicates thru SPLIT_POINTING_ENABLE.  need removing RPC code (keyball, CPI etc. related)
