+++ keyboards/yowkees/keyball/keyball61/config.h
@@ -87,5 +87,8 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 #define SPI_MOSI_PIN B2
 #define SPI_MISO_PIN B3
 
+#define SPLIT_POINTING_ENABLE
+#define POINTING_DEVICE_LEFT
+
 #define POINTING_DEVICE_ROTATION_90
 #define POINTING_DEVICE_INVERT_Y
